move matches component into components folder
