+++ frontend/src/app/app.module.ts
@@ -19,8 +19,8 @@ import { SwipePageComponent } from './components/swipepage/swipepage.component';
 import { AllUsersComponent } from './components/allusers/allusers.component';
 import { UserProfileComponent } from './components/userpage/userpage.component';
 import { AuthInterceptor } from '../services/auth.interceptor';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import { MatchesComponent } from './matches/matches.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MatchesComponent } from './components/matches/matches.component';
 
 @NgModule({
     declarations: [
@@ -43,7 +43,7 @@ import { MatchesComponent } from './matches/matches.component';
         GoogleMapsModule,
         ReactiveFormsModule,
         FormsModule,
-        BrowserAnimationsModule
+        BrowserAnimationsModule,
     ],
     providers: [
         {
